requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.11

# Tools
django-extensions==3.0.5

# Testing
mypy==0.782
pytest==6.0.1
pytest-sugar==0.9.4
pytest-django==3.9.0
factory-boy==2.11.1
faker==4.1.2

# Code quality
flake8==3.6.0
